requirements.txt
paddlepaddle==3.0
numpy==1.23.5
scipy==1.10.1
hydra-core==1.3.2
omegaconf==2.3.0
matplotlib==3.7.1